+++ examples/basic/buttons.py
@@ -4,11 +4,10 @@ from vedo import Plotter, Mesh, dataurl, printc
 
 # Define a function that toggles the transparency of a mesh
 #  and changes the button state
-def buttonfunc(evt):
-    if evt.actor and evt.actor.name == "mybutton":
-        mesh.alpha(1 - mesh.alpha())  # toggle mesh transparency
-        bu.switch()                   # change to next status
-        printc(bu.status(), box="_", dim=True)
+def buttonfunc(obj, ename):
+    mesh.alpha(1 - mesh.alpha())  # toggle mesh transparency
+    bu.switch()                   # change to next status
+    printc(bu.status(), box="_", dim=True)
 
 # Load a mesh and set its color to violet
 mesh = Mesh(dataurl+"magnolia.vtk").c("violet").flat()
@@ -19,15 +18,14 @@ plt = Plotter(axes=11)
 # Add a button to the plotter with buttonfunc as the callback function
 bu = plt.add_button(
     buttonfunc,
-    pos=(0.7, 0.05),  # x,y fraction from bottom left corner
+    pos=(0.7, 0.1),   # x,y fraction from bottom left corner
     states=["click to hide", "click to show"],  # text for each state
     c=["w", "w"],     # font color for each state
     bc=["dg", "dv"],  # background color for each state
     font="courier",   # font type
-    size=25,          # font size
+    size=30,          # font size
     bold=True,        # bold font
     italic=False,     # non-italic font style
-    name="mybutton",
 )
 
 # Show the mesh, docstring, and button in the plot